examples: move vulkan mod out of common mod
Fixes unused imports warnings.
